+++ Items/Tokens/tier3/SapphireToken.cs
@@ -36,34 +36,34 @@ namespace prefixtest.Items.Tokens.tier3
 		switch (x)
 		{
 				case 0:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<vitalitycharm>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<vitalitycharm>(), 1);
 						break;
 				case 1:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<slug>(), 999);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<slug>(), 999);
 						break;
 				case 2:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<brokenengine>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<brokenengine>(), 1);
 						break;
 				case 3:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<ceaselesshunger>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<ceaselesshunger>(), 1);
 						break;
 				case 4:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<GhastlyKalis>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<GhastlyKalis>(), 1);
 						break;
 				case 5:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<hourglass>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<hourglass>(), 1);
 						break;
 				case 6:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<purifierresolve>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<purifierresolve>(), 1);
 						break;
 				case 7:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<seatofcommand>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<seatofcommand>(), 1);
 						break;
 				case 8:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<tuningfork>(), 1);
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<tuningfork>(), 1);
 						break;
 				case 9:
-						Item.NewItem( player.position.X,  player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
+						Item.NewItem( (int)  player.position.X,  (int) player.position.Y, player.width, player.height, ModContent.ItemType<soulofchance>(), Main.rand.Next(2, 5));
 						break;
 		}
 		return true;
